Реализована работа с PlaceTemporaryLiveData - временной LiveData для хранения настроек на экране "Выбор места работы".
