+++ app/src/main/java/ru/practicum/android/diploma/filters/data/impl/FiltersLocalRepositoryImpl.kt
@@ -4,7 +4,7 @@ import android.content.SharedPreferences
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
 import ru.practicum.android.diploma.filters.domain.FiltersLocalRepository
-import ru.practicum.android.diploma.filters.domain.model.FilterParameters
+import ru.practicum.android.diploma.common.presentation.model.FilterParameters
 
 class FiltersLocalRepositoryImpl(
     private val gson: Gson,
