+++ app/src/main/java/ru/practicum/android/diploma/filters/domain/FiltersLocalInteractor.kt
@@ -1,6 +1,6 @@
 package ru.practicum.android.diploma.filters.domain
 
-import ru.practicum.android.diploma.filters.domain.model.FilterParameters
+import ru.practicum.android.diploma.common.presentation.model.FilterParameters
 
 interface FiltersLocalInteractor {
