--- app/src/main/java/ru/practicum/android/diploma/filters/domain/model/FilterParameters.kt
@@ -1,9 +1,0 @@
-package ru.practicum.android.diploma.filters.domain.model
-
-data class FilterParameters(
-    val country: Area?,
-    val region: Area?,
-    val industry: Industry?,
-    val expectedSalary: Int?,
-    val onlyWithSalary: Boolean = false
-)
