+++ app/src/main/java/ru/practicum/android/diploma/filters/presentation/models/PlaceParameters.kt
@@ -1,0 +1,8 @@
+package ru.practicum.android.diploma.filters.presentation.models
+
+import ru.practicum.android.diploma.filters.domain.model.Area
+
+data class PlaceParameters(
+    val countryTemp: Area?,
+    val regionTemp: Area?
+)
